+++ src/services/queries.js
@@ -6,7 +6,7 @@ const { logger } = require('./../libs/logger');
 /**
  *
  * @typedef {object} accountUserName
- * @property {string} username acount username
+ * @property {string} username account username
  */
 /**
  * add new account to database
@@ -17,7 +17,7 @@ async function addAccount({ username }) {
   try {
     const result = await db.accounts.create({ username });
     logger.Info('Queries', 'addAccount', `Add new account: ${JSON.stringify(result)}`);
-    return { accountId: result.id };
+    return { id: result.id };
   } catch (error) {
     if (error.message.includes('E11000 duplicate key')) throw 'username already exists';
     throw error;
@@ -26,8 +26,28 @@ async function addAccount({ username }) {
 
 /**
  *
+ * @typedef {object} accountIDUserName
+ * @property {string} accountId account unique id
+ * @property {string} username account username
+ */
+/**
+ * add new account to database
+ * @param {accountIDUserName} {username}
+ * @return {Promise<AccountID>}
+ */
+async function getAccount({ accountId, username }) {
+  const query = {};
+  if (accountId) query._id = accountId;
+  if (username) query.username = username;
+  const result = await db.accounts.findOne(query);
+  if (!result) throw 'Not found account';
+  return { id: result.id, username: result.username };
+}
+
+/**
+ *
  * @typedef {object} AccountID
- * @property {string } accountId account uuid - ObjectId
+ * @property {string } id account uuid - ObjectId
  */
 /**
  * remove account from database
@@ -89,7 +109,7 @@ async function addProduct({ name, amount }) {
   try {
     const result = await db.products.create({ name, amount });
     logger.Info('Queries', 'addProduct', `add new products: ${JSON.stringify(result)}`);
-    return { productId: result.id };
+    return { id: result.id };
   } catch (error) {
     if (error.message.includes('E11000 duplicate key')) throw 'product collection already exists';
     throw error;
@@ -105,17 +125,22 @@ async function addProduct({ name, amount }) {
 /**
  * get list of products holded by account from database
  * @param {ProductIdAndAmount} {productId,amount}
- * @return {Promise<void>}
+ * @return {Promise<Product>}
  */
 async function updateProductStock({ productId, amount }) {
   const found = await db.products.findOneAndUpdate({ _id: productId }, { $inc: { amount } }, { new: true });
   if (found) {
     logger.Info('Queries', 'updateProductStock', `update product stock: ${JSON.stringify(found)}`);
+    return { id: found.id, name: found.name, amount: found.amount };
   } else {
     throw 'Not found product';
   }
 }
-
+/**
+ * @typedef {object} ProductIDName
+ * @property {string} productId
+ * @property {string} name
+ */
 /**
  * @typedef {object} Product
  * @property {string} id
@@ -123,26 +148,37 @@ async function updateProductStock({ productId, amount }) {
  * @property {number} amount
  */
 /**
- * list all products from database
+ * list all products or based on productId or name, from database
+ * @param {ProductIDName} {productId,name}
  * @return {Promise<Array<Product>}
  */
-async function listProducts() {
-  const result = await db.products.find({});
+async function listProducts({ productId, name }) {
+  const query = {};
+  if (productId) query._id = productId;
+  if (name) query.name = name;
+  const result = await db.products.find(query);
+  if (!result.length) throw 'Empty depot or found product';
   logger.Info('Queries', 'listProducts', `available stock: ${JSON.stringify(result)}`);
-
   return result.map(({ id, name, amount }) => ({ id, name, amount }));
 }
 
 /**
+ * @typedef {object} CartWithoutID
+ * @property {string} accountId
+ * @property {productId} productId
+ * @property {number} amount
+ */
+/**
  * @typedef {object} Cart
+ * @property {string} id
  * @property {string} accountId
  * @property {productId} productId
  * @property {number} amount
  */
 /**
  * add product to a user cart
- * @param {Cart} {accountId,productId,amount}
- * @return {Promise<void>}
+ * @param {CartWithoutID} {accountId,productId,amount}
+ * @return {Promise<Cart>}
  */
 async function holdProduct({ accountId, productId, amount }) {
   const account = await db.accounts.findById({ _id: accountId }).populate({ path: '_accountCart', match: { product: productId } });
@@ -154,9 +190,10 @@ async function holdProduct({ accountId, productId, amount }) {
         { new: true }
       );
       if (product) {
-        await db.carts.create({ holder: accountId, product: productId, amount });
+        const result = await db.carts.create({ holder: accountId, product: productId, amount });
         logger.Info('Queries', 'holdProduct', `account ${accountId} holds ${amount} products ${productId}`);
         logger.Info('Queries', 'holdProduct', `remaining ${product.amount} products ${productId} in stock`);
+        return { id: result.id, holder: result.holder, product: result.product, amount: result.amount };
       } else {
         throw 'Not found product or product remaining stock is not enough';
       }
@@ -171,7 +208,7 @@ async function holdProduct({ accountId, productId, amount }) {
 /**
  * update quandtity of a product hold by an user by increasing or decreasing an amount
  * @param {Cart} {accountId,productId,amount}
- * @return {Promise<void>}
+ * @return {Promise<Cart>}
  */
 async function updateCartAmount({ accountId, productId, amount }) {
   const account = await db.accounts.findById({ _id: accountId }).populate({ path: '_accountCart', match: { product: productId } });
@@ -186,6 +223,7 @@ async function updateCartAmount({ accountId, productId, amount }) {
         const newData = await db.carts.findOneAndUpdate({ holder: accountId, product: productId }, { $inc: { amount } }, { new: true });
         logger.Info('Queries', 'updateCartAmount', `account ${accountId} holds ${newData.amount} products ${productId}`);
         logger.Info('Queries', 'updateCartAmount', `remaining ${product.amount} products of ${product.name} (${productId}) in stock`);
+        return { id: newData.id, holder: newData.holder, product: newData.product, amount: newData.amount };
       } else {
         throw 'Not found product or product remaining stock is not enough';
       }
@@ -194,6 +232,19 @@ async function updateCartAmount({ accountId, productId, amount }) {
     }
   } else {
     throw 'Not found account';
+  }
+}
+
+async function getProductHolders({ productId, name }) {
+  const query = {};
+  if (productId) query._id = productId;
+  if (name) query.name = name;
+  const found = await db.products.findOne(query);
+  if (found) {
+    const result = await db.carts.find({ product: found.id }).populate('_account');
+    return result.map((cart) => cart._account[0]);
+  } else {
+    throw 'Not found product';
   }
 }
 
@@ -206,6 +257,7 @@ async function updateCartAmount({ accountId, productId, amount }) {
 /**
  * move product out from cart and from stock
  * @param {AccountIDProductID} {accountId,productId}
+ * @return {AccountIDProductID} {accountId,productId}
  */
 async function moveCart({ accountId, productId }) {
   const account = await db.accounts.findById({ _id: accountId });
@@ -216,6 +268,7 @@ async function moveCart({ accountId, productId }) {
       if (cart) {
         await db.carts.deleteOne({ holder: accountId, product: productId });
         logger.Info('Queries', 'holdProduct', `account ${accountId} move out holded products ${product.name} ${productId}`);
+        return { id: cart.id, holder: cart.holder, product: cart.product, amount: 0 };
       } else {
         throw `account ${accountId} doesn't hold those kind of product ${productId}`;
       }
@@ -229,12 +282,14 @@ async function moveCart({ accountId, productId }) {
 
 module.exports = {
   addAccount,
+  getAccount,
   removeAccount,
   getAccountHolds,
   addProduct,
   updateProductStock,
   listProducts,
   holdProduct,
+  getProductHolders,
   updateCartAmount,
   moveCart
 };