+++ tests/units/config.test.js
@@ -72,9 +72,12 @@ describe(`Test 'config loading' code`, () => {
     test('it should contains mandatory fields', () => {
       const { server } = config;
       const expected = {
+        apiType: 'graphql',
         protocol: 'http',
         port: 8080,
-        basicAuth: { admin_test: 'admin_test_pwd' },
+        adminLogin: 'admin_test',
+        adminPassword: 'admin_test_pwd',
+        token_secret: 'admin-secret',
         apikey: 'apiKey test 123',
         ssl: { key: undefined, cert: undefined }
       };